Upgrade to .NET 4.
Because I am moving away from RegisterHotkey function. The reason is
that RegisterHotKey only allows one key combination globally. This
means that any other app wanting to use a similar hotkey is not allowed
to do so. As a result, this is preparation for moving towards
WH_KEYBOARD_LL to do the hooking of the keyboard.

Since someone already wrote something really good for WH_KEYBOARD_LL
hooking, it makes no sense for me to rewrite it. As a result, I am
upgrading to .NET 4.0 to be able to utilize their class.