Adding process title tracker.
We can't rely on Process.MainWindowTitle because it doesn't change.
We need to hook into the EVENT_OBJECT_NAMECHANGE event and check
if it is the process that we are interested in. This will be used
later for mintty process to check when the title changes and update
the tab and window accordingly.