Adding hacks to keep children windows focused.
This change hooks into a mouse click on the title bar and resize
event to keep the children window focused. A further hack needed
to be done on the title bar click because we can't intercept the
WH_NCLBUTTONUP event, it gets swallowed up for one reason or
another. A way around this is to detect MOUSEMOVE after a
BUTTONDOWN.