Finally, an alt tab solution that works!
At first I tried the WM_ACTIVATE, didn't work because the event
does not give enough information (or I misused it) to determine
if you are activated from a child or an external window.

Then I tried a hook on EVENT_SYSTEM_FOREGROUND, but for some reason,
the event gets swallowed randomly. So it would sometimes activate
the child and sometimes it wouldn't.

The final solution that did work is a SHELLHOOK to detect when a
window is activated. Using the trick of determining if the activated
window is a child or not and set the focus properly.