Tab title color now changes when there's output.
When there's output in a tab that's not in the foreground, the tab
title changes its color to red. This requires a modification in
DockPanel Suite to allow for a change in the text color and a way
to refresh the content individually when updates were made.

This also required a change in mintty to output its content on stdout
when passing in "-l -" option in the command line. You can follow
the issue at code.google.com/p/mintty/issues/detail?id=337, hopefully
the patch gets accepted. Otherwise, I will provide the mintty version
with that change along with this application.