Fix bugs on initial create and null exception.
For the null exception, it's when the last tab gets closed and the
code tries to focus on ActiveDocument. Since it no longer exists, a
null pointer exception is thrown.

The second bug is on the creation of the first tab. It needs to be
focused after it is created.