Removing another compiler warning.
We were defining the Finalize method, which the compiler gave us a
warning. Not sure how things worked in previous versions of .NET, but
a breakpoint inside the Finalize never got hit. Effectively, I am not
sure if we ever unhooked the WinEvent. I moved the unhook to the
OnHandleDestroyed method, which gets executed when the panel is closed.