Fixed an annoying bug that caused new lines.
This only exists in mintty. When the window gets minimized, the width
gets set to 0. So the child mintty window's width get set to 0. This
somehow caused an "enter key" behavior behavior on the mintty window
and the prompt goes down one line. I don't even know how I found this
bug, it's mainly luck :).