Fixed bug where handle is lost.
When we call OnHandleDestroyed after killing off the child window.
The event for the HandleDestroyed event inside of ctlPuttyPanel tries
to retrieve the child handle and gets 0. As a result, it does not
remove the child handle from frmSupperPutty.children mapping. Since
we rely on the size of this dictionary for selecting tabs, we are now
going off of the end of the tab index. Null pointer :(.