Two fixes, didn't know why I didn't commit earlier
First fix: Remove child panel mapping when we are closing out the
child window. This was insignificant before, but now that we are
relying on the content of this mapping to get a valid ctlPuttyPanel,
we need to prune old instances. Otherwise we will be having problems
with null references.

Second fix: Changing tabs between a putty tab and a mintty tab does
not result in the title being changed. It could have been the case
between mintty windows as well, but I didn't notice it then. Anyway,
this now updates the window's text when the tabs are being changed.