Making sure tab text color gets reset/changed.
There was never a reset for the hasData private variable. This variable
determines if it will have to update the color, and if it does, it will
only happen once when unfocused. The problem was that this variable never
got reset back to false. So the text color only changed once, ever.